+++ src/lib/pages/account-details/index.tsx
@@ -18,6 +18,7 @@ import {
   useInternalNavigate,
   useMoveConfig,
   useNftConfig,
+  useTierConfig,
   useValidateAddress,
   useWasmConfig,
 } from "lib/app-provider";
@@ -30,9 +31,9 @@ import { CelatoneSeo } from "lib/components/Seo";
 import { InvalidState } from "lib/components/state";
 import { UserDocsLink } from "lib/components/UserDocsLink";
 import { useFormatAddresses } from "lib/hooks/useFormatAddresses";
-import { useAccountData } from "lib/services/accountService";
+import { useAccountData } from "lib/services/account";
 import { useModulesByAddress } from "lib/services/move/module";
-import { useResourcesByAddress } from "lib/services/move/resourceService";
+import { useResourcesByAddressLcd } from "lib/services/move/resource";
 import { useNftsCountByAccount } from "lib/services/nft";
 import type { Addr, BechAddr, HexAddr, Option } from "lib/types";
 import { truncate } from "lib/utils";
@@ -79,6 +80,7 @@ const AccountDetailsBody = ({
   const navigate = useInternalNavigate();
   const { address: accountAddress, hex: hexAddress } =
     formatAddresses(accountAddressParam);
+  const isFullTier = useTierConfig() === "full";
 
   // ------------------------------------------//
   // ------------------QUERIES-----------------//
@@ -94,7 +96,7 @@ const AccountDetailsBody = ({
   const { data: modulesData, isFetching: isModulesLoading } =
     useModulesByAddress({ address: accountAddress });
   const { data: resourcesData, isFetching: isResourceLoading } =
-    useResourcesByAddress(accountAddress);
+    useResourcesByAddressLcd(accountAddress);
   // nft
   const { data: nftsCount, isFetching: isNftsCountLoading } =
     useNftsCountByAccount(hexAddress);
@@ -184,7 +186,7 @@ const AccountDetailsBody = ({
             isDisabled={nftsCount === 0}
             onClick={handleTabChange(TabIndex.Nfts, nftsCount)}
             isLoading={isNftsCountLoading}
-            hidden={!nft.enabled}
+            hidden={!nft.enabled || !isFullTier}
           >
             NFTs
           </CustomTab>
@@ -207,7 +209,7 @@ const AccountDetailsBody = ({
               tableCounts.codesCount ?? undefined
             )}
             isLoading={isLoadingAccountTableCounts}
-            hidden={!wasm.enabled}
+            hidden={!wasm.enabled || !isFullTier}
           >
             Codes
           </CustomTab>
@@ -231,7 +233,7 @@ const AccountDetailsBody = ({
               tableCounts.contractsAdminCount ?? undefined
             )}
             isLoading={isLoadingAccountTableCounts}
-            hidden={!wasm.enabled}
+            hidden={!wasm.enabled || !isFullTier}
           >
             Admins
           </CustomTab>
@@ -264,7 +266,7 @@ const AccountDetailsBody = ({
               tableCounts.proposalsCount ?? undefined
             )}
             isLoading={isLoadingAccountTableCounts}
-            hidden={!gov.enabled}
+            hidden={!gov.enabled || !isFullTier}
           >
             Proposals
           </CustomTab>
@@ -331,7 +333,7 @@ const AccountDetailsBody = ({
                 </Flex>
               )}
             </Flex>
-            {nft.enabled && (
+            {nft.enabled && isFullTier && (
               <NftsOverview
                 totalCount={nftsCount}
                 userAddress={hexAddress}
@@ -349,16 +351,18 @@ const AccountDetailsBody = ({
             />
             {wasm.enabled && (
               <>
-                <StoredCodesTable
-                  address={accountAddress}
-                  scrollComponentId={tableHeaderId}
-                  totalData={tableCounts.codesCount ?? undefined}
-                  refetchCount={refetchCounts}
-                  onViewMore={handleTabChange(
-                    TabIndex.Codes,
-                    tableCounts.codesCount ?? undefined
-                  )}
-                />
+                {isFullTier && (
+                  <StoredCodesTable
+                    address={accountAddress}
+                    scrollComponentId={tableHeaderId}
+                    totalData={tableCounts.codesCount ?? undefined}
+                    refetchCount={refetchCounts}
+                    onViewMore={handleTabChange(
+                      TabIndex.Codes,
+                      tableCounts.codesCount ?? undefined
+                    )}
+                  />
+                )}
                 <InstantiatedContractsTable
                   address={accountAddress}
                   scrollComponentId={tableHeaderId}
@@ -369,16 +373,18 @@ const AccountDetailsBody = ({
                     tableCounts.contractsCount ?? undefined
                   )}
                 />
-                <AdminContractsTable
-                  address={accountAddress}
-                  scrollComponentId={tableHeaderId}
-                  totalData={tableCounts.contractsAdminCount ?? undefined}
-                  refetchCount={refetchCounts}
-                  onViewMore={handleTabChange(
-                    TabIndex.Admins,
-                    tableCounts.contractsAdminCount ?? undefined
-                  )}
-                />
+                {isFullTier && (
+                  <AdminContractsTable
+                    address={accountAddress}
+                    scrollComponentId={tableHeaderId}
+                    totalData={tableCounts.contractsAdminCount ?? undefined}
+                    refetchCount={refetchCounts}
+                    onViewMore={handleTabChange(
+                      TabIndex.Admins,
+                      tableCounts.contractsAdminCount ?? undefined
+                    )}
+                  />
+                )}
               </>
             )}
             {move.enabled && (
@@ -402,7 +408,7 @@ const AccountDetailsBody = ({
                 />
               </>
             )}
-            {gov.enabled && (
+            {gov.enabled && isFullTier && (
               <OpenedProposalsTable
                 address={accountAddress}
                 scrollComponentId={tableHeaderId}