+++ src/lib/pages/block-details/full.tsx
@@ -1,0 +1,32 @@
+import { Breadcrumb } from "lib/components/Breadcrumb";
+import { Loading } from "lib/components/Loading";
+import { UserDocsLink } from "lib/components/UserDocsLink";
+import { useBlockData } from "lib/services/block";
+
+import { BlockDetailsTop, BlockInfo, BlockTxsTableFull } from "./components";
+import { InvalidBlock } from "./components/InvalidBlock";
+
+export const BlockDetailsFull = ({ height }: { height: number }) => {
+  const { data: blockData, isLoading } = useBlockData(height);
+
+  if (isLoading) return <Loading withBorder />;
+  if (!blockData) return <InvalidBlock />;
+  return (
+    <>
+      <Breadcrumb
+        items={[
+          { text: "Blocks", href: "/blocks" },
+          { text: blockData.height.toString() },
+        ]}
+      />
+      <BlockDetailsTop blockData={blockData} />
+      <BlockInfo blockData={blockData} />
+      <BlockTxsTableFull height={height} />
+      <UserDocsLink
+        title="What is a block?"
+        cta="Read more about Block"
+        href="general/blocks/detail-page"
+      />
+    </>
+  );
+};