+++ src/lib/pages/migrate/index.tsx
@@ -7,9 +7,11 @@ import { trackToMigrate } from "lib/amplitude";
 import {
   useCurrentChain,
   useInternalNavigate,
+  useTierConfig,
   useWasmConfig,
 } from "lib/app-provider";
 import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
+import { ContractInputSection } from "lib/components/ContractInputSection";
 import { ContractSelectSection } from "lib/components/ContractSelectSection";
 import { CustomIcon } from "lib/components/icon";
 import { FooterCTA } from "lib/components/layouts";
@@ -39,6 +41,7 @@ const Migrate = () => {
   useWasmConfig({ shouldRedirect: true });
   const router = useRouter();
   const navigate = useInternalNavigate();
+  const isFullTier = useTierConfig() === "full";
   const { data: uploadAccessParams, isFetching } = useUploadAccessParamsLcd();
   const {
     proceed,
@@ -182,11 +185,18 @@ const Migrate = () => {
           subtitle="You need to connect your wallet to perform this action"
         />
         {/* Select Migrate Contract modal */}
-        <ContractSelectSection
-          mode="only-admin"
-          contractAddress={contractAddress}
-          onContractSelect={onContractSelect}
-        />
+        {isFullTier ? (
+          <ContractSelectSection
+            mode="only-admin"
+            contractAddress={contractAddress}
+            onContractSelect={onContractSelect}
+          />
+        ) : (
+          <ContractInputSection
+            contract={contractAddress}
+            onContractSelect={onContractSelect}
+          />
+        )}
         <Box mt={12} w="full">
           {renderBody()}
         </Box>
@@ -195,7 +205,7 @@ const Migrate = () => {
         <FooterCTA
           cancelButton={{
             leftIcon: <CustomIcon name="chevron-left" />,
-            onClick: router.back,
+            onClick: handleBack,
           }}
           actionButton={{
             isDisabled: isDisabledProcess,