+++ src/lib/pages/past-txs/lite.tsx
@@ -46,6 +46,7 @@ export const PastTxsLite = () => {
     pageSize,
     offset,
     {
+      enabled: !!address,
       onSuccess: ({ total }) => setTotalData(total),
     }
   );