+++ src/lib/pages/proposal-details/data.ts
@@ -1,7 +1,19 @@
+import { useTierConfig } from "lib/app-provider";
 import { useAssetInfos } from "lib/services/assetService";
 import { useMovePoolInfos } from "lib/services/move/poolService";
-import { useProposalData } from "lib/services/proposal";
-import type { Nullable, Option, ProposalData } from "lib/types";
+import {
+  useProposalData,
+  useProposalDataLcd,
+  useProposalDepositsLcd,
+  useProposalVotesInfo,
+} from "lib/services/proposal";
+import type {
+  Nullable,
+  Option,
+  ProposalData,
+  ProposalVotesInfo,
+} from "lib/types";
+import { ProposalStatus } from "lib/types";
 import { coinToTokenWithValue } from "lib/utils";
 
 interface DerivedProposalDataResponse {
@@ -9,12 +21,39 @@ interface DerivedProposalDataResponse {
     info: Nullable<ProposalData>;
   }>;
   isLoading: boolean;
+  isDepositsLoading: boolean;
 }
 
 export const useDerivedProposalData = (
   id: number
 ): DerivedProposalDataResponse => {
-  const { data, isLoading } = useProposalData(id);
+  const isFullTier = useTierConfig() === "full";
+  const { data: dataApi, isLoading: isApiLoading } = useProposalData(
+    id,
+    isFullTier
+  );
+  const { data: dataLcd, isLoading: isLcdLoading } = useProposalDataLcd(
+    id,
+    !isFullTier
+  );
+  const { data: dataDepositsLcd, isLoading: isDepositsLcdLoading } =
+    useProposalDepositsLcd(id, !isFullTier);
+
+  const [data, isLoading, isDepositsLoading] = isFullTier
+    ? [dataApi, isApiLoading, isApiLoading]
+    : [
+        dataLcd
+          ? {
+              info: {
+                ...dataLcd,
+                proposalDeposits: dataDepositsLcd ?? [],
+              },
+            }
+          : undefined,
+        isLcdLoading,
+        isDepositsLcdLoading,
+      ];
+
   const { data: assetInfos, isLoading: isAssetInfosLoading } = useAssetInfos({
     withPrices: false,
   });
@@ -25,12 +64,15 @@ export const useDerivedProposalData = (
     return {
       data: undefined,
       isLoading: isLoading || isAssetInfosLoading || isMovePoolInfosLoading,
+      isDepositsLoading:
+        isDepositsLoading || isAssetInfosLoading || isMovePoolInfosLoading,
     };
 
   if (!data)
     return {
       data: undefined,
       isLoading: false,
+      isDepositsLoading: false,
     };
 
   if (!data.info)
@@ -39,6 +81,7 @@ export const useDerivedProposalData = (
         info: null,
       },
       isLoading: false,
+      isDepositsLoading: false,
     };
 
   return {
@@ -68,5 +111,33 @@ export const useDerivedProposalData = (
       },
     },
     isLoading: false,
+    isDepositsLoading,
+  };
+};
+
+interface DerivedProposalVotesInfoResponse {
+  data: Option<ProposalVotesInfo>;
+  isLoading: boolean;
+}
+export const useDerivedProposalVotesInfo = (
+  id: number,
+  proposalData: DerivedProposalDataResponse["data"],
+  isProposalDataLoading: boolean
+): DerivedProposalVotesInfoResponse => {
+  const isVotingPeriod =
+    proposalData?.info?.status === ProposalStatus.VOTING_PERIOD;
+
+  const { data, isLoading } = useProposalVotesInfo(id, isVotingPeriod);
+
+  if (!isVotingPeriod) {
+    return {
+      data: proposalData?.info?.finalTallyResult,
+      isLoading: isProposalDataLoading,
+    };
+  }
+
+  return {
+    data,
+    isLoading,
   };
 };