+++ src/lib/services/searchService.ts
@@ -4,10 +4,12 @@ import {
   useCelatoneApp,
   useCurrentChain,
   useGetAddressType,
+  useTierConfig,
   useValidateAddress,
 } from "lib/app-provider";
-import type { BechAddr, Option } from "lib/types";
+import type { Addr, BechAddr, Nullish, Option } from "lib/types";
 import { zBechAddr32, zValidatorAddr } from "lib/types";
+import type { IcnsNamesByAddress } from "lib/types/name";
 import {
   isHexModuleAddress,
   isHexWalletAddress,
@@ -16,14 +18,17 @@ import {
   splitModule,
 } from "lib/utils";
 
-import { useBlockData } from "./block";
+import { useBlockData, useBlockDataLcd } from "./block";
 import { useModuleByAddressLcd } from "./move/module";
-import { useAddressByICNSName, useICNSNamesByAddress } from "./nameService";
-import type { ICNSNamesResponse } from "./ns";
+import { useAddressByIcnsNameLcd, useIcnsNamesByAddressLcd } from "./name";
 import { usePoolByPoolId } from "./poolService";
-import { useProposalData } from "./proposal";
+import { useProposalData, useProposalDataLcd } from "./proposal";
 import { useTxData } from "./tx";
-import { useValidatorData } from "./validator";
+import {
+  useAddressByInitiaUsername,
+  useInitiaUsernameByAddress,
+} from "./username";
+import { useValidatorData, useValidatorDataLcd } from "./validator";
 import { useCodeLcd } from "./wasm/code";
 import { useContractData } from "./wasm/contract";
 
@@ -40,22 +45,30 @@ export type SearchResultType =
 
 export interface ResultMetadata {
   icns: {
-    icnsNames: Option<ICNSNamesResponse>;
+    icnsNames: Option<IcnsNamesByAddress>;
     address: Option<BechAddr>;
     bech32Prefix: string;
   };
+  initiaUsername: {
+    username: Nullish<string>;
+    address: Nullish<Addr>;
+    bech32Prefix: string;
+  };
+}
+
+interface SearchHandlerResponse {
+  results: SearchResultType[];
+  isLoading: boolean;
+  metadata: ResultMetadata;
 }
 
 // eslint-disable-next-line complexity
 export const useSearchHandler = (
   keyword: string,
   resetHandlerStates: () => void
-): {
-  results: SearchResultType[];
-  isLoading: boolean;
-  metadata: ResultMetadata;
   // eslint-disable-next-line sonarjs/cognitive-complexity
-} => {
+): SearchHandlerResponse => {
+  const isFullTier = useTierConfig() === "full";
   const [debouncedKeyword, setDebouncedKeyword] = useState(keyword);
   const {
     chainConfig: {
@@ -98,24 +111,43 @@ export const useSearchHandler = (
 
   // ICNS
   const { data: icnsAddressData, isFetching: icnsAddressFetching } =
-    useAddressByICNSName(debouncedKeyword);
+    useAddressByIcnsNameLcd(debouncedKeyword);
   // provide ICNS metadata result
-  const { data: icnsNames } = useICNSNamesByAddress(
-    isAddr ? (debouncedKeyword as BechAddr) : icnsAddressData?.address
-  );
+  const address = (
+    isAddr ? debouncedKeyword : icnsAddressData?.address
+  ) as BechAddr;
+  const { data: icnsNames } = useIcnsNamesByAddressLcd(address);
+
+  // Initia Username
+  const {
+    data: initiaUsernameAddrData,
+    isFetching: initiaUsernameAddrFetching,
+  } = useAddressByInitiaUsername(debouncedKeyword);
+
+  const { data: initiaUsername, isFetching: initiaUsernameFetching } =
+    useInitiaUsernameByAddress(
+      isAddr ? (debouncedKeyword as Addr) : initiaUsernameAddrData?.address,
+      isMove
+    );
 
   const addressResult = useMemo(() => {
     if (isAddr) {
+      // eslint-disable-next-line sonarjs/no-duplicate-string
       return contractData ? "Contract Address" : "Account Address";
     }
-    return (
-      icnsAddressData?.address &&
-      (icnsAddressData.addressType === "contract_address"
-        ? "Contract Address"
-        : "Account Address")
-    );
-  }, [isAddr, contractData, icnsAddressData]);
 
+    if (icnsAddressData?.address)
+      return (
+        icnsAddressData.address &&
+        (icnsAddressData.addressType === "contract_address"
+          ? "Contract Address"
+          : "Account Address")
+      );
+
+    if (initiaUsernameAddrData?.address) return "Account Address";
+
+    return false;
+  }, [isAddr, icnsAddressData, initiaUsernameAddrData, contractData]);
   // Code
   const { data: codeData, isFetching: codeFetching } = useCodeLcd(
     Number(debouncedKeyword),
@@ -128,28 +160,57 @@ export const useSearchHandler = (
   const { data: txData, isFetching: txFetching } = useTxData(debouncedKeyword);
 
   // Block
-  const { data: blockData, isFetching: blockFetching } = useBlockData(
+  const blockApi = useBlockData(
     Number(debouncedKeyword),
-    isPosDecimal(debouncedKeyword)
+    isPosDecimal(debouncedKeyword) && isFullTier
   );
+  const blockLcd = useBlockDataLcd(
+    Number(debouncedKeyword),
+    isPosDecimal(debouncedKeyword) && !isFullTier
+  );
+  const { data: blockData, isFetching: blockFetching } = isFullTier
+    ? blockApi
+    : blockLcd;
 
   // Proposal
-  const { data: proposalData, isFetching: proposalFetching } = useProposalData(
-    Number(debouncedKeyword),
-    isGov && isId(debouncedKeyword)
-  );
+  const { data: proposalApiData, isFetching: proposalApiIsFetching } =
+    useProposalData(
+      Number(debouncedKeyword),
+      isGov && isId(debouncedKeyword) && isFullTier
+    );
+  const { data: proposalLcdData, isFetching: proposalLcdIsFetching } =
+    useProposalDataLcd(
+      Number(debouncedKeyword),
+      isGov && isId(debouncedKeyword) && !isFullTier
+    );
+  const [proposalData, proposalFetching] = isFullTier
+    ? [proposalApiData, proposalApiIsFetching]
+    : [
+        proposalLcdData
+          ? {
+              info: proposalLcdData,
+            }
+          : undefined,
+        proposalLcdIsFetching,
+      ];
 
   // Validator
-  const { data: validatorData, isFetching: validatorFetching } =
-    useValidatorData(
-      zValidatorAddr.parse(debouncedKeyword),
-      isGov && validateValidatorAddress(debouncedKeyword) === null
-    );
+  const validatorApi = useValidatorData(
+    zValidatorAddr.parse(debouncedKeyword),
+    isGov && validateValidatorAddress(debouncedKeyword) === null && isFullTier
+  );
+  const validatorLcd = useValidatorDataLcd(
+    zValidatorAddr.parse(debouncedKeyword),
+    isGov && validateValidatorAddress(debouncedKeyword) === null && !isFullTier
+  );
+  const { data: validatorData, isFetching: validatorFetching } = isFullTier
+    ? validatorApi
+    : validatorLcd;
 
   // Pool
   const { data: poolData, isFetching: poolFetching } = usePoolByPoolId(
     Number(debouncedKeyword),
-    isPool && isId(debouncedKeyword)
+    isPool && isId(debouncedKeyword) && isFullTier
   );
 
   // Move
@@ -205,15 +266,22 @@ export const useSearchHandler = (
       contractFetching ||
       blockFetching ||
       icnsAddressFetching ||
+      initiaUsernameAddrFetching ||
+      initiaUsernameFetching ||
       poolFetching ||
       proposalFetching ||
       validatorFetching,
     metadata: {
       icns: {
         icnsNames,
+        address,
+        bech32Prefix,
+      },
+      initiaUsername: {
+        username: initiaUsername?.username,
         address: isAddr
-          ? (debouncedKeyword as BechAddr)
-          : icnsAddressData?.address,
+          ? (debouncedKeyword as Addr)
+          : initiaUsernameAddrData?.address,
         bech32Prefix,
       },
     },