+++ src/lib/services/wasm/contract/index.ts
@@ -14,13 +14,7 @@ import type {
   ContractTableCounts,
   MigrationHistoriesResponse,
 } from "lib/services/types";
-import type {
-  BechAddr,
-  BechAddr20,
-  BechAddr32,
-  JsonDataType,
-  Option,
-} from "lib/types";
+import type { BechAddr, BechAddr32, JsonDataType, Option } from "lib/types";
 
 import {
   getAdminContractsByAddress,
@@ -32,6 +26,7 @@ import {
   getMigrationHistoriesByContractAddress,
 } from "./api";
 import {
+  getContractCw2InfoLcd,
   getContractLcd,
   getContractQueryLcd,
   getContractQueryMsgsLcd,
@@ -89,7 +84,8 @@ export const useMigrationHistoriesByContractAddress = (
 };
 
 export const useMigrationHistoriesByContractAddressLcd = (
-  contractAddress: BechAddr32
+  contractAddress: BechAddr32,
+  enabled = true
 ) => {
   const endpoint = useLcdEndpoint();
 
@@ -104,6 +100,7 @@ export const useMigrationHistoriesByContractAddressLcd = (
     {
       retry: 1,
       keepPreviousData: true,
+      enabled,
     }
   );
 };
@@ -247,7 +244,7 @@ export const useContractQueryLcd = (
 };
 
 export const useInstantiatedContractsByAddressLcd = (
-  address: Option<BechAddr20>,
+  address: Option<BechAddr>,
   enabled = false
 ) => {
   const endpoint = useLcdEndpoint();
@@ -266,6 +263,19 @@ export const useInstantiatedContractsByAddressLcd = (
       return getInstantiatedContractsByAddressLcd(endpoint, address);
     },
     { enabled: Boolean(address) && enabled, refetchOnWindowFocus: false }
+  );
+};
+
+export const useContractCw2InfoLcd = (
+  contractAddress: BechAddr32,
+  enabled = true
+) => {
+  const endpoint = useLcdEndpoint();
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.CONTRACT_CW2_INFO_LCD, endpoint, contractAddress],
+    async () => getContractCw2InfoLcd(endpoint, contractAddress),
+    { retry: 1, refetchOnWindowFocus: false, enabled }
   );
 };